// vscode/src/generated-sources/modality-api/models/WorkspaceSegmentId.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Modality REST API
 * Modality REST API
 *
 * The version of the OpenAPI document: 2.0
 * Contact: support@auxon.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * A specific segment of a workspace.
 * @export
 * @interface WorkspaceSegmentId
 */
export interface WorkspaceSegmentId {
    /**
     * 
     * @type {string}
     * @memberof WorkspaceSegmentId
     */
    ruleName: string;
    /**
     * 
     * @type {string}
     * @memberof WorkspaceSegmentId
     */
    segmentName: string;
    /**
     * 
     * @type {string}
     * @memberof WorkspaceSegmentId
     */
    workspaceVersionId: string;
}

/**
 * Check if a given object implements the WorkspaceSegmentId interface.
 */
export function instanceOfWorkspaceSegmentId(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "ruleName" in value;
    isInstance = isInstance && "segmentName" in value;
    isInstance = isInstance && "workspaceVersionId" in value;

    return isInstance;
}

export function WorkspaceSegmentIdFromJSON(json: any): WorkspaceSegmentId {
    return WorkspaceSegmentIdFromJSONTyped(json, false);
}

export function WorkspaceSegmentIdFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkspaceSegmentId {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'ruleName': json['rule_name'],
        'segmentName': json['segment_name'],
        'workspaceVersionId': json['workspace_version_id'],
    };
}

export function WorkspaceSegmentIdToJSON(value?: WorkspaceSegmentId | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'rule_name': value.ruleName,
        'segment_name': value.segmentName,
        'workspace_version_id': value.workspaceVersionId,
    };
}