// vscode/src/generated-sources/modality-api/models/WorkspaceSegmentMetadata.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Modality REST API
 * Modality REST API
 *
 * The version of the OpenAPI document: 2.0
 * Contact: support@auxon.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { WorkspaceSegmentId } from './WorkspaceSegmentId';
import {
    WorkspaceSegmentIdFromJSON,
    WorkspaceSegmentIdFromJSONTyped,
    WorkspaceSegmentIdToJSON,
} from './WorkspaceSegmentId';

/**
 * A specific segment of a workspace.
 * @export
 * @interface WorkspaceSegmentMetadata
 */
export interface WorkspaceSegmentMetadata {
    /**
     * 
     * @type {WorkspaceSegmentId}
     * @memberof WorkspaceSegmentMetadata
     */
    id: WorkspaceSegmentId;
    /**
     * 
     * @type {number}
     * @memberof WorkspaceSegmentMetadata
     */
    latestReceiveTime?: number | null;
}

/**
 * Check if a given object implements the WorkspaceSegmentMetadata interface.
 */
export function instanceOfWorkspaceSegmentMetadata(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;

    return isInstance;
}

export function WorkspaceSegmentMetadataFromJSON(json: any): WorkspaceSegmentMetadata {
    return WorkspaceSegmentMetadataFromJSONTyped(json, false);
}

export function WorkspaceSegmentMetadataFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkspaceSegmentMetadata {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': WorkspaceSegmentIdFromJSON(json['id']),
        'latestReceiveTime': !exists(json, 'latest_receive_time') ? undefined : json['latest_receive_time'],
    };
}

export function WorkspaceSegmentMetadataToJSON(value?: WorkspaceSegmentMetadata | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': WorkspaceSegmentIdToJSON(value.id),
        'latest_receive_time': value.latestReceiveTime,
    };
}